Restructure code to separate out different components
